moving TERMINAL_SIZE() and GET_KBRD() from basic_fun.cpp in terminfo.cpp; trying to add management of Lines and Columns in TERMINAL_SIZE(cols [,lines]). May not work outside Linux+Readline (TBC)
