+++ src/terminfo.hpp
@@ -15,10 +15,25 @@
  *                                                                         *
  ***************************************************************************/
 
+#include "includefirst.hpp"
+
 #ifndef TERMINFO_HPP_
 #define TERMINFO_HPP_
+
+#include "envt.hpp"
+
+#if defined(HAVE_LIBREADLINE)
+void SetTermSize(int rows, int cols);
+#endif
 
 int TermWidth();
 int TermHeight();
 
+namespace lib {
+  using namespace std;
+  
+  BaseGDL* terminal_size_fun( EnvT* e );
+  
+  BaseGDL* get_kbrd( EnvT* e);
+}
 #endif
